requirements.txt
dxh_py==0.0.5
sh==2.0.7; sys_platform != "win32"
binaryornot==0.4.4

# Code quality
# ------------------------------------------------------------------------------
ruff==0.6.2
djlint==1.34.1
black==24.4.2
isort==5.13.2
flake8==7.0.0
django-upgrade==1.20.0
pre-commit==3.8.0

# Testing
# ------------------------------------------------------------------------------
tox==4.18.0
pytest==8.3.2
pytest-xdist==3.6.1
pytest-cookies==0.7.0
pytest-instafail==0.5.0
pyyaml==6.0.2

# Scripting
# ------------------------------------------------------------------------------
PyGithub==2.4.0
gitpython==3.1.43
jinja2==3.1.4
requests==2.32.3